requirements.txt
youtube-transcript-api
yt-dlp
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
requests
pandas